refactor(stats): extract cover routes and limit concurrent backfills
- Move cover image routes into stats-cover-routes.ts
- Cap background anime cover fetches at 3 concurrent tasks
- Add test asserting the concurrency limit
